splitted datamangement in multiple services
